+++ admin-app/frontend/src/views/OrganizationStructure.vue
@@ -54,23 +54,6 @@
                 @click="clickOnEntity(entity)"
               >
                 {{ entity.name }}
-                <br />
-                vertical order
-                {{
-                  verticalOrderByEntityId(
-                    entity.entityId,
-                    hierarchy.hierarchyId
-                  )
-                }}
-                <br />
-                parent is above entity:
-                {{
-                  parentIsAboveEntity(
-                    entity.entityId,
-                    entity.upperEntityId,
-                    entity.hierarchyId
-                  )
-                }}
                 <v-btn fab icon class="entity-icon">
                   <v-icon color="darken-2"> mdi-pencil-outline </v-icon>
                 </v-btn>
@@ -83,122 +66,7 @@
       <div
         class="column dotted-left-border d-flex flex-column align-center justify-center"
       >
-        <v-dialog
-          v-model="newLevelDialogIsDisplayed"
-          max-width="800px"
-          class="dialog"
-          persistent
-        >
-          <template v-slot:activator="{ on, attrs }">
-            <v-btn
-              rounded
-              x-large
-              color="primary"
-              @click="showNewLevelDialog(true)"
-              v-bind="attrs"
-              v-on="on"
-            >
-              <v-icon class="mr-2"> mdi-plus </v-icon>
-              {{ $t("organizationStructure.addNewLevel") }}
-            </v-btn>
-          </template>
-
-          <v-card class="px-4 pt-4">
-            <v-form ref="form" @submit.prevent="submitNewLevel" lazy-validation>
-              <v-card-title> <h2>Create new level</h2> </v-card-title>
-              <v-card-subtitle>
-                This is the description of this beautiful dialog.
-              </v-card-subtitle>
-
-              <v-card-text>
-                <v-container>
-                  <v-row>
-                    <v-col cols="12" sm="6">
-                      <v-card-title> Level information </v-card-title>
-                      <v-text-field
-                        v-model="levelName"
-                        :rules="[rules.required]"
-                        :label="
-                          $t('organizationStructure.newLevelDialog.levelName')
-                        "
-                        required
-                        outlined
-                        dense
-                      ></v-text-field>
-                      <v-textarea
-                        v-model="levelDescription"
-                        :counter="
-                          levelDescription.length > levelDescriptionMaxChar - 20
-                        "
-                        :rules="[rules.maxChar]"
-                        :label="
-                          $t(
-                            'organizationStructure.newLevelDialog.levelDescription'
-                          )
-                        "
-                        required
-                        outlined
-                        dense
-                      ></v-textarea>
-
-                      <v-select
-                        v-model="levelIsSubordinateTo"
-                        :items="hierarchialStructure"
-                        :label="
-                          $t(
-                            'organizationStructure.newLevelDialog.levelIsSubordinateTo'
-                          )
-                        "
-                        dense
-                        outlined
-                        persistent-hint
-                        item-value="hierarchyId"
-                        item-text="name"
-                      ></v-select>
-                    </v-col>
-
-                    <v-col cols="12" sm="6">
-                      <v-card-title> Technologies </v-card-title>
-                      <v-card-text>
-                        <v-select
-                          v-model="levelAllowedTechnologies"
-                          :items="technologies"
-                          :label="
-                            $t(
-                              'organizationStructure.newLevelDialog.manageAllowedTechnologies'
-                            )
-                          "
-                          multiple
-                          dense
-                          outlined
-                          persistent-hint
-                          item-value="technologyId"
-                          item-text="name"
-                        ></v-select>
-                      </v-card-text>
-                    </v-col>
-                  </v-row>
-                </v-container>
-              </v-card-text>
-
-              <v-card-actions>
-                <v-spacer></v-spacer>
-                <v-btn color="primary" text @click="showNewLevelDialog(false)">
-                  {{ $t("general.cancel") }}
-                </v-btn>
-                <v-btn
-                  type="submit"
-                  color="primary"
-                  text
-                  @click.prevent="submitNewLevel()"
-                  :disabled="!levelFormIsInvalid"
-                >
-                  {{ $t("general.save") }}
-                </v-btn>
-              </v-card-actions>
-            </v-form>
-          </v-card>
-        </v-dialog>
+        <NewLevelModal />
       </div>
     </div>
   </div>
@@ -207,31 +75,14 @@
 <script>
 import { mapGetters, mapActions } from "vuex";
 import AddEntityButton from "../components/organizationStructure/AddEntityButton.vue";
-
-const levelDescriptionMaxChar = parseInt(
-  process.env.VUE_APP_LEVEL_DESCRIPTION_MAX_CHAR,
-  10
-);
+import NewLevelModal from "../components/organizationStructure/NewLevelModal.vue";
 
 export default {
   name: "OrganizationStructure",
-  components: { AddEntityButton },
+  components: { AddEntityButton, NewLevelModal },
   data() {
-    return {
-      levelDescriptionMaxChar,
-      newLevelDialogIsDisplayed: true,
-      rules: {
-        required: (value) => !!value || this.requiredi18n,
-        maxChar: (value) =>
-          value.length <= levelDescriptionMaxChar || this.maxCharExceededi18n,
-      },
-      levelName: "",
-      levelDescription: "",
-      levelAllowedTechnologies: [],
-      levelIsSubordinateTo: 0,
-    };
+    return {};
   },
-  mounted() {},
   computed: {
     ...mapGetters({
       allEntitiesOfHierarchy: "entities/getAllEntitiesOfHierarchyByHid",
@@ -251,37 +102,12 @@ export default {
         "entities/getMinVerticalOrderOfTreeRootDescendantsInAHierarchy",
       getEntityShouldHaveVerticalLine:
         "entities/getEntityShouldHaveVerticalLine",
-      technologies: "entities/getTechnologies",
     }),
-    requiredi18n: function () {
-      return this.$t("login.required");
-    },
-    maxCharExceededi18n: function () {
-      return this.$t("login.maxCharExceeded", {
-        maxChar: levelDescriptionMaxChar,
-      });
-    },
-    levelFormIsInvalid: function () {
-      return !!(this.levelName && this.levelDescription);
-    },
   },
   methods: {
     ...mapActions({
       clickOnEntity: "entities/clickOnEntity",
-      saveNewLevel: "entities/saveNewLevel",
     }),
-    showNewLevelDialog: function (payload) {
-      this.newLevelDialogIsDisplayed = payload;
-    },
-    submitNewLevel: function () {
-      this.showNewLevelDialog(false);
-      this.saveNewLevel({
-        levelName: this.levelName,
-        levelDescription: this.levelDescription,
-        technologies: this.levelAllowedTechnologies,
-        upperHierarchy: this.levelIsSubordinateTo,
-      });
-    },
   },
 };
 </script>
@@ -326,9 +152,5 @@ export default {
   position: absolute;
   top: 0;
   right: 0;
-}
-
-.dialog {
-  width: 75%;
 }
 </style>
